complete verify function

+++ include/solver.h
@@ -15,7 +15,7 @@ typedef struct KeccakSolver {
   Options options;
   MathSystem mathSystem;
   uint8_t *device_mq_buffer;
-  uint32_t *device_output_buffer;
+  uint8_t *device_output_buffer;
   threadpool_t *threadpool;
 } KeccakSolver;
